Don't visit "current" symlink when making report.
In fact, this should be the other way around. "current" symlink should
points to the directory that we are looking for, so that should be 
the only directory that we visit.

But let's keep like this for some time, until we're sure that the 
"current" symlink *always* points to the correct directory.